+++ src/plotting/Plotter.cpp
@@ -17,7 +17,7 @@ Plotter::Plotter(string fileName): output(fileName)
 }
 
 // Requires not empty fields, etc.
-void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int> &cornerIndexRow, const Matrix<int> &cornerIndexCol, const Support &supports, const vector<Force> &forces, const int solutionSteps, const float residum, const vector<float> &stress, const float time) 
+void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<size_t> &cornerIndexRow, const Matrix<size_t> &cornerIndexCol, const Support &supports, const vector<Force> &forces, const size_t solutionSteps, const float residum, const vector<float> &stress, const float time) 
 {
     // head
     output << head << endl;
@@ -25,13 +25,13 @@ void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int>
     // field
     output << "const field = [";
     output << "[" << field.Plane(0, 0);
-    for (int c = 1; c < field.Cols; c++)
+    for (size_t c = 1; c < field.Cols; c++)
         output << ", " << field.Plane(0, c);
     output << "]";
-    for (int r = 1; r < field.Rows; r++)
+    for (size_t r = 1; r < field.Rows; r++)
     {
         output << ", [" << field.Plane(r, 0);
-        for (int c = 1; c < field.Cols; c++)
+        for (size_t c = 1; c < field.Cols; c++)
             output << ", " << field.Plane(r, c);
         output << "]";
     }
@@ -39,7 +39,7 @@ void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int>
 
     // u, prepend a leading zero
     output << "const u = [" << 0;
-    for (int i = 0; i < u.size(); i++)
+    for (size_t i = 0; i < u.size(); i++)
         output << ", " << u[i];
     output << "];" << endl;
 
@@ -47,13 +47,13 @@ void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int>
     // cornerIndexRow
     output << "const cornerIndexRow = [";
     output << "[" << cornerIndexRow.Value(0, 0);
-    for (int c = 1; c < cornerIndexRow.getCols(); c++)
+    for (size_t c = 1; c < cornerIndexRow.getCols(); c++)
         output << ", " << cornerIndexRow.Value(0, c);
     output << "]";
-    for (int r = 1; r < cornerIndexRow.getRows(); r++)
+    for (size_t r = 1; r < cornerIndexRow.getRows(); r++)
     {
         output << ", [" << cornerIndexRow.Value(r, 0);
-        for (int c = 1; c < cornerIndexRow.getCols(); c++)
+        for (size_t c = 1; c < cornerIndexRow.getCols(); c++)
             output << ", " << cornerIndexRow.Value(r, c);
         output << "]";
     }
@@ -62,13 +62,13 @@ void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int>
     // cornerIndexCol
     output << "const cornerIndexCol = [";
     output << "[" << cornerIndexCol.Value(0, 0);
-    for (int c = 1; c < cornerIndexCol.getCols(); c++)
+    for (size_t c = 1; c < cornerIndexCol.getCols(); c++)
         output << ", " << cornerIndexCol.Value(0, c);
     output << "]";
-    for (int r = 1; r < cornerIndexCol.getRows(); r++)
+    for (size_t r = 1; r < cornerIndexCol.getRows(); r++)
     {
         output << ", [" << cornerIndexCol.Value(r, 0);
-        for (int c = 1; c < cornerIndexCol.getCols(); c++)
+        for (size_t c = 1; c < cornerIndexCol.getCols(); c++)
             output << ", " << cornerIndexCol.Value(r, c);
         output << "]";
     }
@@ -77,11 +77,11 @@ void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int>
     // supports
     output << "const supports = { RowSupports: [";
     output << "{ row: " << supports.RowSupports[0].row << ", col: " << supports.RowSupports[0].col << " }";
-    for (int i = 1; i < supports.RowSupports.size(); i++)
+    for (size_t i = 1; i < supports.RowSupports.size(); i++)
         output << ", { row: " << supports.RowSupports[i].row << ", col: " << supports.RowSupports[i].col << " }";
     output << "], ColSupports: [";
     output << "{ row: " << supports.ColSupports[0].row << ", col: " << supports.ColSupports[0].col << " }";
-    for (int i = 1; i < supports.ColSupports.size(); i++)
+    for (size_t i = 1; i < supports.ColSupports.size(); i++)
         output << ", { row: " << supports.ColSupports[i].row << ", col: " << supports.ColSupports[i].col << " }";
     output << "]}" << endl;
 
@@ -89,7 +89,7 @@ void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int>
     output << "const forces = [";
     output << "{ attackCorner: { row: " << forces[0].attackCorner.row << ", col: " << forces[0].attackCorner.col << "}, "
             << "forceRow: " << forces[0].forceRow << ", forceCol:" << forces[0].forceCol << "}";
-    for (int i = 1; i < forces.size(); i++)
+    for (size_t i = 1; i < forces.size(); i++)
     {
         output << ", { attackCorner: { row: " << forces[i].attackCorner.row << ", col: " << forces[i].attackCorner.col << "}, "
                 << "forceRow: " << forces[i].forceRow << ", forceCol:" << forces[i].forceCol << "}";
@@ -104,7 +104,7 @@ void Plotter::plot(const Field &field, const vector<float> &u, const Matrix<int>
 
     // stress
     output << "const stress = [" << stress[0];
-    for (int i = 1; i < stress.size(); i++)
+    for (size_t i = 1; i < stress.size(); i++)
         output << ", " << stress[i];
     output << "];" << endl;
